Create Gillespie.cpp
This is a Gillespie model simulator, that takes arbitrary numbers of variables and equations, specified by an input file. (When I started programming it, I didn't do the file approach; will be fixed in next commit.) The simulation should be as fast as possible, but does not parallelize, as most threads would be waiting most of the time anyway (in fact, the synchronization of threads may take longer than the speedup you might get from parallelization). It will use two arrays, one to store the number of instances of each variable (1D array of length var), and one to store the equations and their rates (2D array of size equations * (var + 1)).